Moved methods to individual source files.
