+++ src/lib/include/utilities.h
@@ -1,5 +1,5 @@
-#ifndef CURLY_TYPES_H
-#define CURLY_TYPES_H
+#ifndef CURLY_UTILITIES_H
+#define CURLY_UTILITIES_H
 
 /// \brief Returns the name of a type
 ///
@@ -10,4 +10,4 @@
 /// \return char*
 char* ltype_name(int t);
 
-#endif  /// CURLY_TYPES_H
+#endif  /// CURLY_UTILITIES_H
