+++ src/lib/src/lenv.c
@@ -6,3 +6,98 @@ struct lenv
     char** syms;
     lval** vals;
 };
+
+
+///////////////////////////
+/// `lenv` Constructors ///
+///////////////////////////
+
+lenv* lenv_new(void)
+{
+    lenv* e = malloc(sizeof(lenv));
+    e->count = 0;
+    e->syms = NULL;
+    e->vals = NULL;
+    return e;
+}
+
+
+//////////////////////////
+/// `lenv` Destructors ///
+//////////////////////////
+
+void lenv_del(lenv* e)
+{
+    for (int i = 0; i < e->count; i++)
+    {
+        free(e->syms[i]);
+        lval_del(e->vals[i]);
+    }
+
+    free(e->syms);
+    free(e->vals);
+    free(e);
+}
+
+
+//////////////////////
+/// `lenv` Methods ///
+//////////////////////
+
+lval* lenv_get(lenv* e, lval* k)
+{
+    for (int i = 0; i < e->count; i++)
+        if (strcmp(e->syms[i], k->sym) == 0)
+            return lval_copy(e->vals[i]);
+    
+    return lval_err("Unbound symbol! '%s'", k->sym);
+}
+
+
+void lenv_put(lenv* e, lval* k, lval* v)
+{
+    for (int i = 0; i < e->count; i++)
+        if (strcmp(e->syms[i], k->sym) == 0)
+        {
+            lval_del(e->vals[i]);
+            e->vals[i] = lval_copy(v);
+            return;
+        }
+
+    e->count++;
+    e->vals = realloc(e->vals, sizeof(lval*) * e->count);
+    e->syms = realloc(e->syms, sizeof(char*) * e->count);
+
+    e->vals[e->count - 1] = lval_copy(v);
+    e->syms[e->count - 1] = malloc(strlen(k->sym) + 1);
+    strcpy(e->syms[e->count - 1], k->sym);
+}
+
+
+void lenv_add_builtin(lenv* e, char* name, lbuiltin func)
+{
+    lval* k = lval_sym(name);
+    lval* v = lval_fun(func);
+    lenv_put(e, k, v);
+    lval_del(k);
+    lval_del(v);
+}
+
+
+void lenv_add_builtins(lenv* e)
+{
+    lenv_add_builtin(e, "\\", builtin_lambda);
+
+    lenv_add_builtin(e, "list", builtin_list);
+    lenv_add_builtin(e, "head", builtin_head);
+    lenv_add_builtin(e, "tail", builtin_tail);
+    lenv_add_builtin(e, "eval", builtin_eval);
+    lenv_add_builtin(e, "join", builtin_join);
+
+    lenv_add_builtin(e, "+", builtin_add);
+    lenv_add_builtin(e, "-", builtin_sub);
+    lenv_add_builtin(e, "*", builtin_mul);
+    lenv_add_builtin(e, "/", builtin_div);
+
+    lenv_add_builtin(e, "def", builtin_def);
+}
