+++ src/lib/src/windows.c
@@ -1,4 +1,5 @@
 #include <windows.h>
+#include <stdlib.h>
 
 #ifdef _WIN32
